requirements.txt
fastapi
uvicorn
together
numpy
faiss-cpu
faker
python-dotenv